+++ lib/rspec_jsonapi_serializer/matchers/base.rb
@@ -13,26 +13,16 @@ module RSpecJSONAPISerializer
       end
 
       def failure_message
-        ([main_failure_message] + submatcher_failure_messages).compact.join("\n")
+        raise NotImplementedError
       end
 
       def failure_message_when_negated
-        ([main_failure_message_when_negated] + submatcher_failure_messages_when_negated)
-          .compact
-          .join("\n")
+        raise NotImplementedError
       end
 
       protected
 
       attr_reader :expected, :serializer_instance, :submatchers
-
-      def main_failure_message
-        raise NotImplementedError
-      end
-
-      def main_failure_message_when_negated
-        raise NotImplementedError
-      end
 
       def add_submatcher(submatcher)
         submatchers << submatcher
@@ -48,16 +38,6 @@ module RSpecJSONAPISerializer
 
       def serializer_name
         serializer_instance.class.name
-      end
-
-      private
-
-      def submatcher_failure_messages
-        failing_submatchers.map(&:failure_message)
-      end
-
-      def submatcher_failure_messages_when_negated
-        failing_submatchers.map(&:failure_message_when_negated)
       end
 
       def failing_submatchers