+++ CHANGELOG.md
@@ -4,9 +4,19 @@
 
 - Add `.id_method_name` submatcher for relationship matchers (`belong_to`, `have_one` and `have_many`)
 
+## [1.1.1] - 2023-05-20
+
+- Add default description for relationship serializer submatcher
+
 ## [1.1.0] - 2021-04-15
 
 - Add `.serializer` submatcher for relationship matchers (`belong_to`, `have_one` and `have_many`)
+
+## [1.0.1] - 2023-05-20
+
+- Add default description for have_attribute matcher and submatchers
+- Add default description for have_link matcher and submatchers
+- Add default description for have_meta matcher and submatchers
 
 ## [1.0.0] - 2021-04-10